[Refactor] Move subgraph store into workflow store
